Refactored event system, replacing it all with functionality from Dianode\Events.
The functionality present in Dianode\Events (working title) is a development of that
which was present in BedRest up until now. It has been moved out of BedRest proper
as it is a fully re-usable component in its own right, the only dependency it has is
Doctrine Common.

The composer.json has been updated with the temporary location of this module for
CI and build purposes.